refactor: move all equivocation work in one file
